0.1.3+rev115
- added C++ mod acpi
- fixed C mod acpi

+++ include/acpi.h
@@ -2,4 +2,4 @@
 
 #include "types.h"
 
-u32 ACPI_CoreID();
+u16 ACPI_CoreID();
